+++ config.lua.dist
@@ -176,9 +176,9 @@ location = "Sweden"
 -- NOTE: The API communication uses raw, insecure HTTP
 -- DO NOT SEND private data over it, unless you use an external HTTPS server as a reverse proxy
 -- WARNING: If you change this to an external IP you do it at your own responsibility
-httpApiIp = "127.0.0.1"
-httpApiPort = 7174
+httpIp = "127.0.0.1"
+httpPort = 7174
 
 -- UNIX socket example (socket path: /run/tfs.sock) - not supported on Windows
 -- This is recommended on systems which support local stream sockets (Linux/BSD)
--- httpApiIp = "unix:///run/tfs.sock"
+-- httpIp = "unix:///run/tfs.sock"
