+++ src/CMakeLists.txt
@@ -1,11 +1,11 @@
 set(tfs_SRC
 	${CMAKE_CURRENT_LIST_DIR}/otpch.cpp
 	${CMAKE_CURRENT_LIST_DIR}/actions.cpp
-	${CMAKE_CURRENT_LIST_DIR}/http_api/peer.cpp
-	${CMAKE_CURRENT_LIST_DIR}/http_api/responder.cpp
-	${CMAKE_CURRENT_LIST_DIR}/http_api/router.cpp
-	${CMAKE_CURRENT_LIST_DIR}/http_api/server.cpp
-	${CMAKE_CURRENT_LIST_DIR}/http_api/generic_stream_acceptor.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/http/peer.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/http/responder.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/http/router.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/http/server.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/http/generic_stream_acceptor.cpp
 	${CMAKE_CURRENT_LIST_DIR}/ban.cpp
 	${CMAKE_CURRENT_LIST_DIR}/baseevents.cpp
 	${CMAKE_CURRENT_LIST_DIR}/bed.cpp
