+++ src/configmanager.cpp
@@ -179,7 +179,7 @@ bool ConfigManager::load()
 		string[MYSQL_PASS] = getGlobalString(L, "mysqlPass", "");
 		string[MYSQL_DB] = getGlobalString(L, "mysqlDatabase", "forgottenserver");
 		string[MYSQL_SOCK] = getGlobalString(L, "mysqlSock", "");
-		string[HTTP_API_IP] = getGlobalString(L, "httpApiIp", "127.0.0.1");
+		string[HTTP_IP] = getGlobalString(L, "httpIp", "127.0.0.1");
 
 		integer[SQL_PORT] = getGlobalNumber(L, "mysqlPort", 3306);
 
@@ -194,7 +194,7 @@ bool ConfigManager::load()
 		integer[STATUS_PORT] = getGlobalNumber(L, "statusProtocolPort", 7171);
 
 		integer[MARKET_OFFER_DURATION] = getGlobalNumber(L, "marketOfferDuration", 30 * 24 * 60 * 60);
-		integer[HTTP_API_PORT] = getGlobalNumber(L, "httpApiPort", 7174);
+		integer[HTTP_PORT] = getGlobalNumber(L, "httpPort", 7174);
 	}
 
 	boolean[ALLOW_CHANGEOUTFIT] = getGlobalBoolean(L, "allowChangeOutfit", true);
