+++ src/http/peer.h
@@ -1,14 +1,14 @@
 // Copyright 2022 The Forgotten Server Authors. All rights reserved.
 // Use of this source code is governed by the GPL-2.0 License that can be found in the LICENSE file.
 
-#ifndef FS_HTTP_API_PEER_H
-#define FS_HTTP_API_PEER_H
+#ifndef FS_HTTP_PEER_H
+#define FS_HTTP_PEER_H
 
 #include "../tools.h"
 #include "common.h"
 #include <boost/beast/http.hpp>
 
-namespace HttpApi
+namespace Http
 {
 
 class Server;
@@ -138,6 +138,6 @@ public:
 
 using PeerWeakPtr = std::weak_ptr<Peer>; ///Weak pointer to peer, used to prevent Lua from keeping Peers alive for too long
 
-} //namespace HttpApi
+} //namespace Http
 
-#endif // FS_HTTP_API_PEER_H
+#endif // FS_HTTP_PEER_H
