+++ vc17/theforgottenserver.vcxproj
@@ -185,11 +185,11 @@
     <ClCompile Include="..\src\guild.cpp" />
     <ClCompile Include="..\src\house.cpp" />
     <ClCompile Include="..\src\housetile.cpp" />
-    <ClCompile Include="..\src\http_api\peer.cpp" />
-    <ClCompile Include="..\src\http_api\responder.cpp" />
-    <ClCompile Include="..\src\http_api\router.cpp" />
-    <ClCompile Include="..\src\http_api\generic_stream_acceptor.cpp" />
-    <ClCompile Include="..\src\http_api\server.cpp">
+    <ClCompile Include="..\src\http\peer.cpp" />
+    <ClCompile Include="..\src\http\responder.cpp" />
+    <ClCompile Include="..\src\http\router.cpp" />
+    <ClCompile Include="..\src\http\generic_stream_acceptor.cpp" />
+    <ClCompile Include="..\src\http\server.cpp">
       <ObjectFileName>$(IntDir)\obj_d\api_server</ObjectFileName>
     </ClCompile>
     <ClCompile Include="..\src\inbox.cpp" />
@@ -281,12 +281,12 @@
     <ClInclude Include="..\src\guild.h" />
     <ClInclude Include="..\src\house.h" />
     <ClInclude Include="..\src\housetile.h" />
-    <ClInclude Include="..\src\http_api\common.h" />
-    <ClInclude Include="..\src\http_api\peer.h" />
-    <ClInclude Include="..\src\http_api\responder.h" />
-    <ClInclude Include="..\src\http_api\router.h" />
-    <ClInclude Include="..\src\http_api\generic_stream_acceptor.h" />
-    <ClInclude Include="..\src\http_api\server.h" />
+    <ClInclude Include="..\src\http\common.h" />
+    <ClInclude Include="..\src\http\peer.h" />
+    <ClInclude Include="..\src\http\responder.h" />
+    <ClInclude Include="..\src\http\router.h" />
+    <ClInclude Include="..\src\http\generic_stream_acceptor.h" />
+    <ClInclude Include="..\src\http\server.h" />
     <ClInclude Include="..\src\inbox.h" />
     <ClInclude Include="..\src\iologindata.h" />
     <ClInclude Include="..\src\iomap.h" />
